New feature and update README
Allow to bypass restricted contents in chat where is not allowed.
There are channels where is not allowed capture screen or save
attachments, while is easy to do it in tg desktop and web, on app
is not possible, with this new feature disabled by default, is
possible to save attachments and screen capture.
Screen capture remain prohibited on payment activity always, so
none can capture your credit card.